+++ tests/Integration/Application/Bus/DivideQueryHandler.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Tests\Integration\Application\Bus;
 
-use CloudCreativity\Modules\Application\Messages\Through;
+use CloudCreativity\Modules\Toolkit\Pipeline\Through;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 
 #[Through('division-modifier')]