+++ tests/Unit/Infrastructure/OutboundEventBus/ClosurePublisherTest.php
@@ -13,55 +13,40 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Tests\Unit\Infrastructure\OutboundEventBus;
 
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\IntegrationEvent;
-use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer;
+use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer as IPipeContainer;
 use CloudCreativity\Modules\Infrastructure\OutboundEventBus\ClosurePublisher;
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
 
 class ClosurePublisherTest extends TestCase
 {
-    private MockObject&PipeContainer $middleware;
-
     /**
      * @var array<IntegrationEvent>
      */
     private array $actual = [];
 
-    private ClosurePublisher $publisher;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->publisher = new ClosurePublisher(
-            function (IntegrationEvent $event): void {
-                $this->actual[] = $event;
-            },
-            $this->middleware = $this->createMock(PipeContainer::class),
-        );
-    }
-
     protected function tearDown(): void
     {
         parent::tearDown();
-        unset($this->publisher, $this->middleware, $this->actual);
+        unset($this->actual);
     }
 
     public function test(): void
     {
-        $event = $this->createMock(IntegrationEvent::class);
+        $event = $this->createStub(IntegrationEvent::class);
 
-        $this->publisher->publish($event);
+        $publisher = $this->createPublisher();
+        $publisher->publish($event);
 
         $this->assertSame([$event], $this->actual);
     }
 
     public function testWithMiddleware(): void
     {
-        $event1 = $this->createMock(IntegrationEvent::class);
-        $event2 = $this->createMock(IntegrationEvent::class);
-        $event3 = $this->createMock(IntegrationEvent::class);
-        $event4 = $this->createMock(IntegrationEvent::class);
+        $event1 = $this->createStub(IntegrationEvent::class);
+        $event2 = $this->createStub(IntegrationEvent::class);
+        $event3 = $this->createStub(IntegrationEvent::class);
+        $event4 = $this->createStub(IntegrationEvent::class);
 
         $middleware1 = function ($event, \Closure $next) use ($event1, $event2) {
             $this->assertSame($event1, $event);
@@ -78,44 +63,92 @@ class ClosurePublisherTest extends TestCase
             return $next($event4);
         };
 
-        $this->middleware
+        $middleware = $this->createMock(IPipeContainer::class);
+        $middleware
             ->expects($this->once())
             ->method('get')
             ->with('MySecondMiddleware')
             ->willReturn($middleware2);
 
-        $this->publisher->through([
+        $publisher = $this->createPublisher($middleware);
+        $publisher->through([
             $middleware1,
             'MySecondMiddleware',
             $middleware3,
         ]);
 
-        $this->publisher->publish($event1);
+        $publisher->publish($event1);
 
         $this->assertSame([$event4], $this->actual);
+    }
+
+    public function testWithMiddlewareViaPsrContainer(): void
+    {
+        $event1 = $this->createStub(IntegrationEvent::class);
+        $event2 = $this->createStub(IntegrationEvent::class);
+        $event3 = $this->createStub(IntegrationEvent::class);
+
+        $middleware1 = function ($event, \Closure $next) use ($event1, $event2) {
+            $this->assertSame($event1, $event);
+            return $next($event2);
+        };
+
+        $middleware2 = function ($event, \Closure $next) use ($event2, $event3) {
+            $this->assertSame($event2, $event);
+            return $next($event3);
+        };
+
+        $psrContainer = $this->createMock(ContainerInterface::class);
+        $psrContainer
+            ->expects($this->once())
+            ->method('get')
+            ->with('MySecondMiddleware')
+            ->willReturn($middleware2);
+
+        $publisher = $this->createPublisher($psrContainer);
+        $publisher->through([
+            $middleware1,
+            'MySecondMiddleware',
+        ]);
+
+        $publisher->publish($event1);
+
+        $this->assertSame([$event3], $this->actual);
     }
 
 
     public function testWithAlternativeHandlers(): void
     {
         $expected = new TestOutboundEvent();
-        $mock = $this->createMock(IntegrationEvent::class);
+        $stub = $this->createStub(IntegrationEvent::class);
         $actual = null;
 
-        $this->publisher->bind($mock::class, function (): never {
+        $publisher = $this->createPublisher();
+
+        $publisher->bind($stub::class, function (): never {
             $this->fail('Not expecting this closure to be called.');
         });
 
-        $this->publisher->bind(
+        $publisher->bind(
             TestOutboundEvent::class,
             function (TestOutboundEvent $in) use (&$actual) {
                 $actual = $in;
             },
         );
 
-        $this->publisher->publish($expected);
+        $publisher->publish($expected);
 
         $this->assertEmpty($this->actual);
         $this->assertSame($expected, $actual);
+    }
+
+    private function createPublisher(ContainerInterface|IPipeContainer|null $middleware = null): ClosurePublisher
+    {
+        return new ClosurePublisher(
+            function (IntegrationEvent $event): void {
+                $this->actual[] = $event;
+            },
+            $middleware,
+        );
     }
 }